A content type can no more be deleted if used (#1885)

+++ ApiBundle/Controller/ContentTypeController.php
@@ -85,10 +85,12 @@ class ContentTypeController extends BaseController
      */
     public function deleteAction($contentTypeId)
     {
-        $contentTypes = $this->get('open_orchestra_model.repository.content_type')->findBy(array('contentTypeId' => $contentTypeId));
-        $this->get('open_orchestra_backoffice.manager.content_type')->delete($contentTypes);
-        $this->dispatchEvent(ContentTypeEvents::CONTENT_TYPE_DELETE, new ContentTypeEvent(current($contentTypes)));
-        $this->get('object_manager')->flush();
+        if (0 == $this->get('open_orchestra_model.repository.content')->countByContentType($contentTypeId)) {
+            $contentTypes = $this->get('open_orchestra_model.repository.content_type')->findBy(array('contentTypeId' => $contentTypeId));
+            $this->get('open_orchestra_backoffice.manager.content_type')->delete($contentTypes);
+            $this->dispatchEvent(ContentTypeEvents::CONTENT_TYPE_DELETE, new ContentTypeEvent(current($contentTypes)));
+            $this->get('object_manager')->flush();
+        }
 
         return array();
     }
